requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
nltk==3.8.1
google-api-python-client==2.108.0
wordcloud==1.9.3
matplotlib==3.8.2
requests==2.31.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
